BREAKING: chg default HIGLightGray -> WebLight
This changes the ThemeContext default to WebLight to be inline with the
legacy theme styling that many HIG consumers start with.

If you would like to continue using the HIG Light Gray theme now that
it's no longer the default, you will need to specify that theme when you
initialize the `ThemeContext.Provider`. For instance

```
import HIGLightTheme from "@hig/theme-data/build/json/lightGrayMediumDensityTheme/theme.json";

// ...

<ThemeContext.Provider value={HIGLightTheme}>
  <ThemeContext.Consumer>
    // ...
  </ThemeContext.Consumer>
</ThemeContext.Provider>
```