BREAKING: use in semantic elements Typography
If there is a semantic element that matches the variant provided, use
that semantic element instead of resolving all `<Typography>` components
to a single type of element. Also, this changes the default element type
from `<span>` to `<p>`.

Migrating from 1.0.0 to 2.0.0

If you need the same behavior that the default `<span>` provided before,
you can pass in custom styles to achieve that. For instance,
`<Typography style={{display: "inline"}}>Render nicely</Typography>`.