feat: enable passing elementType to Typography
If you want to enable specifying the semantic element to be rendered by
the component, use the elementType prop. If this prop is not provided,
the semantic element will default to matching the variant (e.g. if the
component variant is "h1", the element will be `<h1>`) or if there is no
semantic element matching the variant, it will default to `<p>`.