feat: make Slider themable
* use all available theme data to enable web light, light gray, dark
gray, dark gray, and dark blue themes
* update snapshot tests
* add theme knob to storybook
* update gemini tests

BREAKING CHANGE:

To enable different styling for the value portion of the range track,
visually distinguishing it from the unfilled portion of the track,
we always set an initial value for a Slider if a value is not explicitly
provided. This initial value default is the midpoint between the min and
max props. THe midpoint will be 50 unless you set min or max. This
behavior mimics the browser behavior for range inputs.