feat: allow active Tab to be controlled
We want to be able to control the active tab from a higher
component. Add activeTabIndex prop to the Tabs component, which can be
used with onTabChange to achieve this control.
Also, add defaultActiveTabIndex to allow the original active tab to be
set when using the Tabs in an uncontrolled manor.