refactor: Major Release - React 17 Upgrade
BREAKING CHANGE: This release includes upgrading to React 17 and all associated libraries. The components have also had structural changes, utilizing stateless components and hooks. There should be no change in look or behavior of components. The code usage is the same so if you're already on react 17 you can bump the version directly. If you're on an old version of react you'll need to upgrade your project's react first to 17 and then the HIG components. This upgrade also means no more fixes for the react 15 version but it will still be available for download from NPM. You can fork the repo and make fixes with the older version if there is something critical past this release date.